Force the generic Java Client API language version to 1.13.2-beta so that we can be sure that our generated code for Cloud Endpoints will be compatible with the client libraries that the codegen endpoint gives us (which has the side effect of making GPE compatible with App Engine 1.7.6, due to a method signature change in the CloudClientLibGenerator class).
Also specify the language version for the 'Add Google APIs' feature - otherwise, there's a risk that we could add a
Google API with a different language version than what the project is using for Cloud Endpoints.

This solution means that we have to manually update the language version in code if users want to get the later version. We should
probably elevate the API language version to a first-class concept so that users can specify it (or at least be aware of it).

Also fixed another incompatibility with App Engine 1.7.6; there was a change to the ServiceContext.create(...) method.
Added code so that GPE will be able to work wih the different versions of this method in App Engine 1.7.5 - App Engine 1.7.7.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=44181201