Fix up problems with client library generation and Cloud Endpoints.
We now annotate the @Api with ownerDomain and packagePath information, so that
a default package structure is not chosen for the user by codegen.

We also annotate method names with specific names, so default naming is not
chosen for the user by the Cloud Endpoints API document generator.

Both of these changes require increasing the minimum version of App Engine
for Cloud Endpoints up to 1.7.7. As such, there was some reflective code
that I was able to remove.

Also did some minor cleanup and refactoring.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=45788214