Use Jigsaw to track unique users importing Apps Script projects and unique projects imported.
I haven't written any sawzall tests because the processing is so trivial that there is nothing meaningful to test. I will not be able to test this end-to-end until the .szl files are checked in and there is some real usage data.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=47382187