A flurry of cleanup/fixes:
1) Generate Endpoints Class action now has reporting and cancellation
2) Errors that occur when generating API metadata as part of the CloudEndpointBuilder are now surfaced via a project-level problem marker (this error is configurable)
3) Removed hacks in the Endpoints sample that deal with versions of App Engine prior to 1.8.0
4) Modified the Java 1.6 deprecation validator to check for Java 1.7 instead (App Engine requires Java 1.7+)

We no longer re-generate the client libraries when the app id changes from the project properties page and the App Engine project is not connected to an Android Project. This feature is of dubious value, and would have led to a bunch of convoluted code to get the progress reporting right. There's no point in doing this since we're moving towards Cloud Endpoints projects with WTP.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=57688630