Merge the com.google.gdt.eclipse.gph.e36 plug-in into com.google.gdt.eclipse.gph, and remove the previous "manual" installation page.
A bit of indirection was added to the installation part of the
Google Project Hosting (gph) plug-in to protect against changes
to the P2 API and to fall back to a page that gives manual
installation instructions.  The P2 API has remained stable and
there is no need for this flexibility, so merge the version-specific
part back into the main gph plug-in, and remove the manual
installation fallback.

Most of the merging involved making P2InstallManager a concrete class
and merging P2InstallManagerImpl into it.

In addtion, the following cleanups have been done:
* Renamed ProjectHostingUIPlugin createStatus() to createErrorStatus()
* Removed the "Install Manager" extension point
* Fixed warnings (added @Overrides and fixed Javadoc)
* Cleaned up and alphabetized MANIFEST.MF files
* Applied auto formatting (first used the “/*-“ trick to keep the
copyright header unchanged)
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=74862366