Merge the internal com.google.gdt.eclipse.suite.ext and ext.e38 plug-ins and move them to the opensource tree as com.google.gdt.eclipse.suite.update. Leave internal Google-specific stuff behind in the process. In particular: * Don't carry the deprecated UpdateSitePing class forward * Don't migrate the OauthClientInfo test extension. I made an attempt to migrate it into a testing class but found that this test has not been running in the legacy build and has likely been broken a long time. I'll follow up with another CL to address that.
As part of the renaming of the plug-in to suite.update, rearrange
the packages so that update-related classes are in the base package,
usage logging-related packages are in a new "usage" subpackage,
and the classes related to importing are in a new "importer" subpackage.

Rajeev, all of these migrated classes seem fine to open source, but
I'd like your feedback on that.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=86014600