Extract the appengine-tools-api.jar file into its own com.google.appengine.eclipse.dependencies plug-in, and make it an optional dependency of the com.google.appengine.eclipse.core plug-in, so that it can be used by Maven/Tycho as a compilation-only dependency.
This rolls back the change to Bundle-Classpath that broke setting
up a new SDK.

This works fine for running automated tests and normal deployment, but
messes up running from the PDE, since by default the PDE launchers will use
all plug-ins from the development environment at runtime. Added logic so that
if running from the PDE is sensed, it will search the running environment
for com.google.appengine.eclipse.dependencies and report how to exclude it.

An alternative to this approach is to do some classloader black magic so
that the parent’s classloader is searched only after all of the child’s
paths, but the documentation around this is rather scary, claiming that
classloaders were not designed to behave this way and there may be
unintended consequences.

Removed support for Appengine SDKs prior to 1.4.3, which was released in
March 2011, nearly 4 years ago. Those prior versions require using the
bundled tools API jar to create the bridge, but now the tools API jar is
a compile-only dependency, not available in the production repository.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=86056613